+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   or `dc_jsonrpc_instance_t` is unreferenced
   during handling the JSON-RPC request. #4153
 - Delete expired messages using multiple SQL requests. #4158
+- Do not emit "Failed to run incremental vacuum" warnings on success. #4160
 
 
 ## 1.111.0